requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
alembic==1.13.2
annotated-types==0.7.0
anyio==4.3.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
databases==0.7.0
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
ormar==0.20.2
pandas==2.2.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.6.1
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
pytz==2024.1
regex==2024.5.15
requests==2.32.3
schedule==1.2.2
six==1.16.0
SQLAlchemy==1.4.54
telegram==0.0.1
twilio==9.5.1
uvicorn==0.34.0
fastapi==0.115.12 
starlette==0.46.1